requirements-dev.txt
annotated-types==0.7.0
APScheduler==3.11.0
astroid==3.3.9
beautifulsoup4==4.13.3
black==25.1.0
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dill==0.3.9
firebase-admin==6.7.0
Flask==3.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-generativeai==0.8.4
Google-Images-Search==1.4.7
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
msgpack==1.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pyfiglet==0.8.post1
PyJWT==2.10.1
pylint==3.3.6
pyparsing==3.2.3
python-resize-image==1.1.20
requests==2.32.3
rsa==4.9
six==1.17.0
soupsieve==2.6
termcolor==1.1.0
tomlkit==0.13.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wikipedia==1.4.0
